Split the round-trip test into its own package, one file per use case
Moves to tests/ndc/roundtrip/ as package roundtrip. Since the framework moves
with the cases, everything stays package-private and nothing has to be exported;
the local package is no longer called ndc either, so the historyndc import alias
goes away.

Cases are now one file per use case, so adding coverage for a component means
adding a file rather than growing a single list:

  framework_test.go              fixture, active driver, artifact, passive apply
  capture_test.go                passive task capture
  diff_test.go                   normalization, allowlist, reporting
  cases_common_test.go           step sequences shared between case files
  cases_workflow_test.go         start, close, backoff-delayed start
  cases_workflowtask_test.go     workflow task lifecycle
  cases_activity_test.go         activity lifecycle and retries
  cases_usertimer_test.go        user timers
  cases_childworkflow_test.go    child workflows

No behavior change: the same eight tests run and pass. Nexus has no case yet, so
it has no file yet -- an empty placeholder would only be noise.
