+++ tests/ndc/roundtrip/capture_test.go
@@ -1,4 +1,4 @@
-package ndc
+package roundtrip
 
 // Task capture for the round trip.
 //
